Move subcommand logic into modules Fixes #17
